bfloat16

+++ cgadimpl/Tests/test_complex_mlp.cpp
@@ -26,7 +26,7 @@ int main() {
 
     // One-hot labels Y[B,Out]
     Tensor Yt(Shape{{B, Out}}, TensorOptions());
-    float* yt_data = Yt.data<float>(); // Get data pointer to fill
+    bfloat16_t* yt_data = Yt.data<bfloat16_t>(); // Get data pointer to fill
     std::mt19937 gen(42);
     std::uniform_int_distribution<int> pick(0, Out - 1);
     for (int i = 0; i < B; ++i) {
@@ -76,15 +76,15 @@ int main() {
 
     // ---------- Report ----------
     // To get a scalar value, move to CPU and get the data pointer
-    float loss_val = loss.val().to_cpu().data<float>()[0];
+    bfloat16_t loss_val = loss.val().to_cpu().data<bfloat16_t>()[0];
     std::cout << "loss = " << loss_val << "\n";
 
     // Show a few logits + softmax probs for the first row
     Value probs = softmax_row(logits);
     Tensor probs_cpu = probs.val().to_cpu();
     Tensor logits_cpu = logits.val().to_cpu();
-    const float* probs_data = probs_cpu.data<float>();
-    const float* logits_data = logits_cpu.data<float>();
+    const bfloat16_t* probs_data = probs_cpu.data<bfloat16_t>();
+    const bfloat16_t* logits_data = logits_cpu.data<bfloat16_t>();
 
     std::cout << "logits[0,:5] = ";
     for (int j = 0; j < std::min(5, Out); ++j)
